Support for GoCmd, add and fix tests, restructure Go commands
