+++ jfrog-cli/jfrog/buildtools_test.go
@@ -3,7 +3,7 @@ package main
 import (
 	"encoding/json"
 	"fmt"
-	"github.com/jfrog/gocmd/golang/project/dependencies"
+	"github.com/jfrog/gocmd/dependencies"
 	gofrogcmd "github.com/jfrog/gofrog/io"
 	"github.com/jfrog/jfrog-cli-go/jfrog-cli/artifactory/commands/generic"
 	"github.com/jfrog/jfrog-cli-go/jfrog-cli/artifactory/commands/gradle"
@@ -347,7 +347,7 @@ func TestGoFallback(t *testing.T) {
 		t.Error(err)
 	}
 
-	err = artifactoryCli.Exec("go", "build", tests.GoLocalRepo, "--recursive-tidy-overwrite=true")
+	err = artifactoryCli.Exec("go", "build", tests.GoLocalRepo)
 	if err != nil {
 		log.Warn(err)
 		if !strings.Contains(err.Error(), dependencies.FailedToRetrieve) || !strings.Contains(err.Error(), dependencies.FromBothArtifactoryAndVcs) {
@@ -371,20 +371,21 @@ func TestGoFallback(t *testing.T) {
 // 3. Build with recursive-tidy-overwrite set to true so the gofrog dependency will be downloaded from VCS
 // 4. Check mod file (in Artifactory) of the dependency gofrog that populated.
 // 5. Check mod file (in Artifactory) of the gofrog dependency (pkg/errors) that exists with the right content
-func TestGoRecursiveTidyOverwrite(t *testing.T) {
+func TestGoRecursivePublish(t *testing.T) {
 	initGoTest(t)
-	testsdataTarget := filepath.Join(tests.Out, "testsdata")
-	testsdataSrc := filepath.Join(tests.GetTestResourcesPath(), "go", "testsdata")
-	err := fileutils.CopyDir(testsdataSrc, testsdataTarget, true)
-	if err != nil {
-		t.Error(err)
-	}
 	wd, err := os.Getwd()
 	if err != nil {
 		t.Error(err)
 	}
 	gopath := os.Getenv("GOPATH")
 	os.Setenv("GOPATH", filepath.Join(wd, tests.Out))
+
+	testsdataTarget := filepath.Join(tests.Out, "testsdata")
+	testsdataSrc := filepath.Join(tests.GetTestResourcesPath(), "go", "testsdata")
+	err = fileutils.CopyDir(testsdataSrc, testsdataTarget, true)
+	if err != nil {
+		t.Error(err)
+	}
 	project1Path := createGoProject(t, "dependency")
 	projectMissingDependency := createGoProject(t, "projectmissingdependency")
 	projectBuild := createGoProject(t, "projectbuild")
@@ -396,27 +397,36 @@ func TestGoRecursiveTidyOverwrite(t *testing.T) {
 	if err != nil {
 		t.Error(err)
 	}
-	err = artifactoryCli.Exec("go", "build", tests.GoLocalRepo, "--recursive-tidy-overwrite=true")
+	err = artifactoryCli.Exec("grp", tests.GoLocalRepo)
 	if err != nil {
 		t.Error(err)
 	}
+	sumFileExists, err := fileutils.IsFileExists("go.sum", false)
+	if err != nil {
+		t.Error(err)
+	}
+	if sumFileExists {
+		err = os.Remove("go.sum")
+		if err != nil {
+			t.Error(err)
+		}
+	}
 	cleanGoCache(t)
-
 	// Need to check the mod file within Artifactory of the gofrog dependency.
 	content := downloadModFile(tests.DownloadModFileGo, wd, "gofrog", t)
 	// Check that the file was signed:
-	if !strings.Contains(string(content), "// Edited by JFrog CLI on") {
+	if !strings.Contains(string(content), "// Generated by JFrog") {
 		t.Error(fmt.Sprintf("Expected file to be signed, however, the file is not signed: %s", string(content)))
 	}
 	// Check that the mod file was populated with the dependency
-	if !strings.Contains(string(content), "require github.com/pkg/errors") {
-		t.Error(fmt.Sprintf("Expected to get one dependency github.com/pkg/errors, however, got: %s", string(content)))
+	if strings.Contains(string(content), "require github.com/pkg/errors") {
+		t.Error(fmt.Sprintf("Expected to get empty mod file, however, got: %s", string(content)))
 	}
 
 	// Need to check the mod file within Artifactory of the dependency of gofrog => pkg/errors.
 	content = downloadModFile(tests.DownloadModOfDependencyGo, wd, "errors", t)
 	// Check that the file was signed:
-	if !strings.Contains(string(content), "// Edited by JFrog CLI on") {
+	if !strings.Contains(string(content), "// Generated by JFrog") {
 		t.Error(fmt.Sprintf("Expected file to be signed, however, the file is not signed: %s", string(content)))
 	}
 	// Check that the mod file was populated with the dependency
